Screen Update
Form1.cs->
"Add Timespan 'avrageTimeToRunDrawing' that is logged in 'DrawCells()' to better debug draw lag specifically"
"Move section of code in FormUpdate that itterates drawing cells to new method 'DrawCells()' for cleaner code. "
"Add handling for nex MpaGen button being pushed and for individual cells being changed to non-mat when clicked and shown. This is to have Map Editor functionality "